requirements.txt
Django>=4.2.0
django-admin-interface
Pillow
python-decouple
django-crispy-forms
crispy-bootstrap5
django-extensions
